+++ crates/oxc_ecmascript/src/lib.rs
@@ -4,8 +4,9 @@ mod bound_names;
 mod is_simple_parameter_list;
 mod private_bound_identifiers;
 mod prop_name;
+mod to_int_32;
 
 pub use self::{
     bound_names::BoundNames, is_simple_parameter_list::IsSimpleParameterList,
-    private_bound_identifiers::PrivateBoundIdentifiers, prop_name::PropName,
+    private_bound_identifiers::PrivateBoundIdentifiers, prop_name::PropName, to_int_32::ToInt32,
 };